meet/requirements.txt
Django==4.2.5
gunicorn==21.2.0
psycopg2-binary==2.9.7 # for postgresql
black==23.9.1
djangorestframework==3.14.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4
drf_yasg==1.21.7
